backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
marshmallow==3.21.1
mysql-connector-python==8.3.0
